% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borders_download.R
\name{borders_download}
\alias{borders_download}
\title{Download State Register of Borders}
\usage{
borders_download(type, outdir = ".", unzip = TRUE, ...)
}
\arguments{
\item{type}{"administrative units", "special units" or "addresses"}

\item{outdir}{(optional) name of the output directory;
by default, files are saved in the working directory}

\item{unzip}{TRUE (default) or FALSE, when TRUE the downloaded archive will
be extracted and removed}

\item{...}{additional argument for \code{\link[utils:download.file]{utils::download.file()}}}
}
\value{
a selected data type in SHP format
}
\description{
Download State Register of Borders
}
\examples{
\dontrun{
borders_download("administrative units") # 366 MB
}
}
