% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commune_names.R
\docType{data}
\name{commune_names}
\alias{commune_names}
\title{Communes in Poland}
\format{
An object of class \code{data.frame} with 2477 rows and 2 columns.
}
\usage{
commune_names
}
\description{
The data frame contains names of communes,
and their identifiers (TERC, 7 characters).
}
\details{
Last update: 2 January 2024
}
\examples{
commune_names
}
\keyword{commune}
\keyword{dataset}
