% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_names.R
\docType{data}
\name{county_names}
\alias{county_names}
\title{Counties in Poland}
\format{
An object of class \code{data.frame} with 380 rows and 3 columns.
}
\usage{
county_names
}
\description{
The data frame contains the names of counties,
their identifiers (TERYT, 4 characters) and the availability of
building models in the LOD2 standard (logical value).
}
\examples{
county_names
}
\keyword{county}
\keyword{dataset}
