% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egib_layer_download.R
\name{egib_layer_download}
\alias{egib_layer_download}
\title{Download Land and Building Register (EGiB) layers}
\usage{
egib_layer_download(
  county = NULL,
  TERYT = NULL,
  layer = "dzialki",
  outdir = NULL,
  ...
)
}
\arguments{
\item{county}{County name in Polish. Check \code{\link[=county_names]{county_names()}} function.}

\item{TERYT}{County ID (4 characters).}

\item{layer}{Requested layer, \code{dzialki} (parcels)default. Other common layer is \code{budynki} (buildings).
You can check available layers by \code{\link[=egib_layers]{egib_layers()}} data set.}

\item{outdir}{(Optional) name of the output directory where the data has to be stored;
NULL by default.}

\item{...}{any other parameter passed to \code{\link[sf:st_read]{sf::st_read()}} function}
}
\value{
Simple feature data frame of objects for requested layer.
}
\description{
Download Land and Building Register (EGiB) layers
}
\examples{
\dontrun{
parcels <- egib_layer_download(county = "wołowski", layer = "dzialki")
buildings <- egib_layer_download(county = "milicki", layer = "budynki", outdir = ".")
}

}
\references{
The EGiB data (Ewidencja Gruntów i Budynków) consist of 2 primary layers:
cadastral data (often called dzialki) and buildings footprints (budynki).
The data is maintained on county level, which results in 380 units/different
sources of data. It may contain additional layers like points of detailed
horizontal and vertical geodetic control network (osnowa_pozioma
and osnowa_pionowa respectively).

\url{https://www.geoportal.gov.pl/en/data/land-and-building-register-egib/}
\url{https://www.geoportal.gov.pl/en/data/detailed-control-network-database-bdsog/}
}
