% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuia_download.R
\name{emuia_download}
\alias{emuia_download}
\title{downloads Register of Towns, Streets and Addresses for communes}
\usage{
emuia_download(commune = NULL, TERYT = NULL, outdir = ".", unzip = TRUE, ...)
}
\arguments{
\item{commune}{commune name in Polish. Check \code{\link[=commune_names]{commune_names()}} function.}

\item{TERYT}{county ID (7 characters)}

\item{outdir}{(optional) name of the output directory;
by default, files are saved in the working directory}

\item{unzip}{TRUE (default) or FALSE, when TRUE the downloaded archive will
be extracted and removed}

\item{...}{additional argument for \code{\link[utils:download.file]{utils::download.file()}}}
}
\value{
a register in SHP format
}
\description{
downloads Register of Towns, Streets and Addresses for communes
}
\examples{
\dontrun{
emuia_download(commune = "Kotla") # 38 KB
emuia_download(TERYT = c("0203042", "2412032")) # 75 KB
}
}
