% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{returns a list with metadata and coordinates for a given type of input
(geocoding addresses and objects)}
\usage{
geocode(address = NULL, road = NULL, rail_crossing = NULL, geoname = NULL)
}
\arguments{
\item{address}{place with or without street and house number}

\item{road}{road number with or without mileage}

\item{rail_crossing}{rail crossing identifier
(11 characters including 2 spaces, format: "XXX XXX XXX")}

\item{geoname}{name of the geographical object from State Register
of Geographical Names (function geonames_download)}
}
\value{
a list(s) with metadata and coordinates (EPSG: 2180)
}
\description{
returns a list with metadata and coordinates for a given type of input
(geocoding addresses and objects)
}
\examples{
geocode(address = "Marki") # place
geocode(address = "Marki, Andersa") # place and street
geocode(address = "Marki, Andersa 1") # place, street and house number
geocode(address = "KrĂłlewskie Brzeziny 13") # place and house number

geocode(road = "632") # road number
geocode(road = "632 55") # road number and mileage

geocode(rail_crossing = "001 018 478")

geocode(geoname = "Las Mierzei") # physiographic object
}
