% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodb_download.R
\name{geodb_download}
\alias{geodb_download}
\title{downloads General Geographic Databases for entire voivodeships}
\usage{
geodb_download(voivodeships, outdir = ".", unzip = TRUE, ...)
}
\arguments{
\item{voivodeships}{selected voivodeships in Polish or English, or TERC
(function 'voivodeship_names' can by helpful)}

\item{outdir}{(optional) name of the output directory}

\item{unzip}{TRUE (default) or FALSE, when TRUE the downloaded archive will be extracted and removed}

\item{...}{additional argument for \code{\link[utils:download.file]{utils::download.file()}}}
}
\value{
a database in Geography Markup Language format (.GML),
the content and detail level corresponds to the general
geographic map in the scale of 1:250000
}
\description{
downloads General Geographic Databases for entire voivodeships
}
\examples{
\donttest{
geodb_download(c("mazowieckie", "wielkopolskie"))
geodb_download(c("Subcarpathian", "Silesian"))
geodb_download(c("02", "16"))

library("sf")
# load roads and territory from "opolskie" (TERC 16)
roads = read_sf("PL.PZGiK.201.16/BDOO/PL.PZGIK.201.16__OT_SKDR_L.xml")
territory = read_sf("PL.PZGiK.201.16/BDOO/PL.PZGIK.201.16__OT_ADJA_A.xml")

# plot them together
plot(territory$geometry)
plot(roads$geometry, add = TRUE, col = "red")
}
}
\references{
description of topographical and general geographical databases,
and technical standards for making maps (in Polish):
\url{http://www.gugik.gov.pl/__data/assets/pdf_file/0005/208661/rozp_BDOT10k_BDOO.pdf}

brief description of categories and layer names (in English and Polish):
\url{https://github.com/kadyb/rgugik/blob/master/vignettes/articles/geodb_description.md}
}
