% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geonames_download.R
\name{geonames_download}
\alias{geonames_download}
\title{downloads State Register of Geographical Names}
\usage{
geonames_download(type, format, ...)
}
\arguments{
\item{type}{names of places ("place") and/or physiographic objects ("object")}

\item{format}{data format ("GML", "SHP" and/or "XLSX")}

\item{...}{additional argument for \code{\link[utils:download.file]{utils::download.file()}}}
}
\value{
a selected data type in the specified format
}
\description{
downloads State Register of Geographical Names
}
\examples{
geonames_download(type = c("place", "object"), format = "SHP")
geonames_download(type = "object", format = c("GML", "XSLX"))
}
\references{
http://isap.sejm.gov.pl/isap.nsf/download.xsp/WDU20150000219/O/D20150219.pdf
}
