% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models3D_download.R
\name{models3D_download}
\alias{models3D_download}
\title{downloads 3D models of buildings for counties}
\usage{
models3D_download(county = NULL, TERYT = NULL, LOD = "LOD1", ...)
}
\arguments{
\item{county}{county name in Polish}

\item{TERYT}{county ID (4 characters)}

\item{LOD}{level of detail for building models ("LOD1" or "LOD2").
"LOD1" is default. "LOD2" is only available for ten voivodeships
(TERC: "04", "06", "12", "14", "16", "18", "20", "24", "26", "28").
Check 'voivodeships_names' function.}

\item{...}{additional argument for \code{\link[utils:download.file]{utils::download.file()}}}
}
\value{
models of buildings in Geography Markup Language format (.GML)
}
\description{
downloads 3D models of buildings for counties
}
\examples{
\donttest{
models3D_download(county = "Toruń")
models3D_download(TERYT = c("2462", "0401"), LOD = "LOD2")
}
}
