% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orto_download.R
\name{orto_download}
\alias{orto_download}
\title{downloads orthoimages based on the data frame obtained using
the orto_request() function}
\usage{
orto_download(df_req, check_SHA = FALSE, ...)
}
\arguments{
\item{df_req}{a data frame obtained using the orto_request() function}

\item{check_SHA}{check the integrity of downloaded files
(logical parameter, default = FALSE)}

\item{...}{}
}
\value{
georeferenced orthoimages .tif with properties
(resolution, composition, etc.) as specified in the input data frame
}
\description{
downloads orthoimages based on the data frame obtained using
the orto_request() function
}
\examples{
library(sf)
polygon_path = system.file("datasets/search_area.gpkg", package = "rgugik")
polygon = st_read(polygon_path)
req_df = orto_request(polygon)
orto_download(req_df[1, ]) # download the first image only
}
