% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointDTM100_download.R
\name{pointDTM100_download}
\alias{pointDTM100_download}
\title{downloads Digital Terrain Models with 100 m resolution for entire voivodeships}
\usage{
pointDTM100_download(voivodeships, ...)
}
\arguments{
\item{voivodeships}{selected voivodeships in Polish or English, or TERC
(function 'voivodeships_names' can by helpful)}

\item{...}{additional argument for \code{\link[utils:download.file]{utils::download.file()}}}
}
\value{
text files with X, Y, Z columns (EPSG:2180)
}
\description{
downloads Digital Terrain Models with 100 m resolution for entire voivodeships
}
\examples{
pointDTM100_download(c("mazowieckie", "wielkopolskie"))
pointDTM100_download(c("Subcarpathian", "Silesian"))
pointDTM100_download(c("02", "16"))
}
