% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointDTM_100_download.R
\name{pointDTM_100_download}
\alias{pointDTM_100_download}
\title{downloads Digital Terrain Models with 100 m resolution for entire voivodeships}
\usage{
pointDTM_100_download(voivodeships, ...)
}
\arguments{
\item{voivodeships}{selected voivodeships in Polish or English, or TERC
(function 'voivodeships_names' can by helpful)}

\item{...}{}
}
\value{
text files with X, Y, Z columns (EPSG:2180)
}
\description{
downloads Digital Terrain Models with 100 m resolution for entire voivodeships
}
\examples{
pointDTM_100_download(c("mazowieckie", "wielkopolskie"))
pointDTM_100_download(c("Subcarpathian", "Silesian"))
pointDTM_100_download(c("02", "16"))
}
