% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgugik-package.R
\docType{package}
\name{rgugik-package}
\alias{rgugik}
\alias{rgugik-package}
\title{rgugik: Search and Retrieve Spatial Data from GUGiK}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kadyb/rgugik}
  \item Report bugs at \url{https://github.com/kadyb/rgugik/issues}
}

}
\author{
\strong{Maintainer}: Krzysztof Dyba \email{adres7@gmail.com} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})

Authors:
\itemize{
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})
}

Other contributors:
\itemize{
  \item GUGiK (source of the data) [contributor]
}

}
\keyword{internal}
