% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TERYT_county.R
\docType{data}
\name{TERYT_county}
\alias{TERYT_county}
\title{toDo}
\format{
An object of class \code{data.frame} with 380 rows and 3 columns.
}
\usage{
TERYT_county
}
\description{
toDo
}
\examples{
data(TERYT_county)
}
\keyword{datasets}
