% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile_download.R
\name{tile_download}
\alias{tile_download}
\title{downloads tiles based on the data frame obtained using
the orto_request() and DEM_request() functions}
\usage{
tile_download(df_req, outdir = ".", check_SHA = FALSE, ...)
}
\arguments{
\item{df_req}{a data frame obtained using the orto_request() and
DEM_request() functions}

\item{outdir}{(optional) name of the output directory}

\item{check_SHA}{check the integrity of downloaded files
(logical parameter, default = FALSE)}

\item{...}{additional argument for \code{\link[utils:download.file]{utils::download.file()}}}
}
\value{
georeferenced tiles with properties (resolution, year, etc.)
as specified in the input data frame
}
\description{
downloads tiles based on the data frame obtained using
the orto_request() and DEM_request() functions
}
\examples{
\donttest{
library(sf)
polygon_path = system.file("datasets/search_area.gpkg", package = "rgugik")
polygon = read_sf(polygon_path)

req_df = orto_request(polygon)
tile_download(req_df[1, ]) # download the first image only

req_df = DEM_request(polygon)
tile_download(req_df[1, ]) # download the first DEM only
}
}
