% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voivodeship_names.R
\docType{data}
\name{voivodeship_names}
\alias{voivodeship_names}
\title{Voivodeships in Poland}
\format{
An object of class \code{data.frame} with 16 rows and 3 columns.
}
\usage{
voivodeship_names
}
\description{
The data frame contains Polish and English names of
voivodeships, and their identifiers (TERC, 2 characters).
}
\examples{
voivodeship_names
}
\keyword{dataset}
\keyword{voivodeship}
