% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voivodeships_names.R
\name{voivodeships_names}
\alias{voivodeships_names}
\title{returns a data frame with Polish and English names of voivodeships and their TERC}
\usage{
voivodeships_names()
}
\value{
a data frame with Polish and English names of voivodeships and their TERC
TERC - territorial division units (voivodeship level, 2 digits)
}
\description{
returns a data frame with Polish and English names of voivodeships and their TERC
}
\examples{
df_names = voivodeships_names()
}
