% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egib_layers.R
\docType{data}
\name{egib_layers}
\alias{egib_layers}
\title{Available URLs and layers of Land and Building Register (EGiB)}
\format{
An object of class \code{data.frame} with 380 rows and 4 columns.
}
\usage{
egib_layers
}
\description{
The data frame contains the names of counties,
their identifiers (TERYT, 4 characters), the URL to the EGiB data (as Web
Feature Service) and the names of available layers.
}
\examples{
egib_layers
}
\keyword{EGiB}
\keyword{dataset}
