% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geonames_download.R
\name{geonames_download}
\alias{geonames_download}
\title{downloads State Register of Geographical Names}
\usage{
geonames_download(type, outdir = ".", unzip = TRUE, format = "SHP", ...)
}
\arguments{
\item{type}{names of places ("place") and/or physiographic objects ("object")}

\item{outdir}{(optional) name of the output directory}

\item{unzip}{TRUE (default) or FALSE, when TRUE the downloaded archive will be extracted and removed}

\item{format}{data format ("GML", "SHP" (default) and/or "XLSX")}

\item{...}{additional argument for \code{\link[utils:download.file]{utils::download.file()}}}
}
\value{
a selected data type in the specified format
}
\description{
downloads State Register of Geographical Names
}
\examples{
\dontrun{
geonames_download(type = c("place", "object"), format = "SHP")
geonames_download(type = "object", format = c("GML", "XSLX"))
}
}
\references{
\url{http://isap.sejm.gov.pl/isap.nsf/download.xsp/WDU20150000219/O/D20150219.pdf}
}
