% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcel_get.R
\name{parcel_get}
\alias{parcel_get}
\title{returns a simple feature geometry for a given TERYT (ID) or coordinates}
\usage{
parcel_get(TERYT = NULL, X = NULL, Y = NULL)
}
\arguments{
\item{TERYT}{parcel ID (18 characters, e.g. "141201_1.0001.6509")}

\item{X}{longitude (EPSG: 2180)}

\item{Y}{latitude (EPSG: 2180)}
}
\value{
a simple feature geometry (in case of TERYT) or data frame with simple
feature geometry and TERYT (in case of coordinates)
}
\description{
returns a simple feature geometry for a given TERYT (ID) or coordinates
}
\examples{
\dontrun{
parcel = parcel_get(TERYT = "141201_1.0001.6509")
parcel = parcel_get(X = 313380.5, Y = 460166.4)
}
}
