% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgugik-package.R
\docType{package}
\name{rgugik-package}
\alias{rgugik}
\alias{rgugik-package}
\title{rgugik: Search and Retrieve Spatial Data from 'GUGiK'}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Automatic open data acquisition from resources of Polish Head Office
    of Geodesy and Cartography ('Główny Urząd Geodezji i Kartografii')
    (<www.gugik.gov.pl>).
    Available datasets include various types of numeric, raster and vector data,
    such as orthophotomaps, digital elevation models (digital terrain models,
    digital surface model, point clouds), state register of borders, spatial
    databases, geometries of cadastral parcels, 3D models of buildings, and more.
    It is also possible to geocode addresses or objects using the geocodePL_get()
    function.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kadyb.github.io/rgugik/}
  \item \url{https://github.com/kadyb/rgugik}
  \item Report bugs at \url{https://github.com/kadyb/rgugik/issues}
}

}
\author{
\strong{Maintainer}: Krzysztof Dyba \email{adres7@gmail.com} (\href{https://orcid.org/0000-0002-8614-3816}{ORCID})

Authors:
\itemize{
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})
}

Other contributors:
\itemize{
  \item Maciej Beręsewicz \email{maciej.beresewicz@ue.poznan.pl} (\href{https://orcid.org/0000-0002-8281-4301}{ORCID}) [contributor]
  \item GUGiK (source of the data) [contributor]
}

}
\keyword{internal}
